refactor(composer): extract shared useOpenPrLink hook
Address review feedback: the openPrLink handler was duplicated verbatim
between Sidebar and the composer branch selector. Extract it into a
shared lib/openPullRequestLink hook and use it from both call sites.
